MidiSetting only exporting component to fix HMR
